create TypeScript definitions for GraphQL responses
